Initial conversion to grails 3
