+++ .gitignore
@@ -1,14 +1,11 @@
-/bin
-/target
-/target-eclipse
-/.settings
-/.classpath
-/.project
-**Thumbs.db
-*.sql
+Thumbs.db
 .DS_Store
-*.log
-/grails-database-migration-*.zip
-/grails-database-migration-*.zip.sha1
-/plugin.xml
-/web-app/WEB-INF
+.gradle
+build/
+.idea
+*.iml
+*.ipr
+*.iws
+.project
+.settings
+.classpath
