--- application.properties
@@ -1,2 +1,0 @@
-app.grails.version=2.3.9
-app.name=slack-logger
