+++ src/main/groovy/slacklogger/SlackloggerGrailsPlugin.groovy
@@ -1,6 +1,9 @@
-class SlackLoggerGrailsPlugin {
-	def version = "1.0.2"
-	def grailsVersion = "2.3 > *"
+package slacklogger
+
+import grails.plugins.*
+
+class SlackLoggerGrailsPlugin extends Plugin {
+	def grailsVersion = "3.0.0 > *"
 	def pluginExcludes = [
 		"grails-app/controllers/*",
 		"grails-app/views/*",
@@ -9,7 +12,8 @@ class SlackLoggerGrailsPlugin {
 	def title = "Slack Logger"
 	def author = "TheConnMan"
 	def authorEmail = "brian@theconnman.com"
-	def description = "Slack Logger is a custom log4j appender which logs directly to a Slack channel."
+	def description = 'Slack Logger is a custom log4j appender which logs directly to a Slack channel.'
+	def profiles = ['web']
 	def documentation = "https://github.com/TheConnMan/SlackLogger"
 	def license = "MIT"
 	def issueManagement = [ system: "GitHub", url: "https://github.com/TheConnMan/SlackLogger/issues" ]
